refactor API to reduce changes for users
